refactor: move bridgeProvider creation to factory
